shell-ui: Convert Alerts library to a set of hooks and component federable
