+++ shell-ui/src/alerts/alertContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const AlertContext = createContext<null>(null);
