+++ shell-ui/src/alerts/index.js
@@ -1,1 +1,2 @@
-import "./library";
+export * from './AlertProvider'
+export * from './alertHooks'
